+++ client/src/admin/SelectOptions.tsx
@@ -1,0 +1,21 @@
+
+export interface CategorysProps {
+    name: string,
+    id: string,
+}
+
+export const Category : CategorysProps[] = [
+    {
+      name: '우울증',
+      id: 'depression',
+    },
+    {
+      name: '불면증',
+      id: 'asomnia',
+    },    
+    {
+      name: '조울증',
+      id: 'tristimania',
+    },
+    
+];
