+++ client/src/admin/components/UI/PageWrapper.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components"
+
+
+
+export const PageWrapper = styled.div`
+    display: grid;
+    grid-template-columns: 200px 1fr;
+    background-color: gray;
+    
+
+`
